# frontend/openapi.json
{"openapi":"3.0.3","info":{"title":"backend","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login to backend","description":"Login to backend\n\nOn successful login you will retrieve a cookie.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful"},"400":{"description":"Client error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}},"/api/v1/auth/logout":{"get":{"tags":["Authentication"],"summary":"Log out of this session","description":"Log out of this session\n\nLogs a logged-in user out of his session.","operationId":"logout","responses":{"200":{"description":"Logout successful"},"400":{"description":"Client error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}},"/api/v1/images/{uuid}":{"get":{"tags":["Search"],"operationId":"get_product_images","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Product image","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductImages"}}}},"400":{"description":"Client error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[{"api_key":[]}]}},"/api/v1/search":{"get":{"tags":["Search"],"operationId":"post_search","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Search results","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchOutput"}}}},"400":{"description":"Client error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[{"api_key":[]}]}}},"components":{"schemas":{"ApiErrorResponse":{"type":"object","description":"The Response that is returned in case of an error\n\nFor client errors the HTTP status code will be 400,\nfor server errors the 500 will be used.","required":["message","status_code"],"properties":{"message":{"type":"string","example":"Error message is here"},"status_code":{"$ref":"#/components/schemas/ApiStatusCode"}}},"ApiStatusCode":{"type":"integer","description":"The status code represents a unique identifier for an error.\n\nError codes in the range of 1000..2000 represent client errors\nthat could be handled by the client.\nError codes in the range of 2000..3000 represent server errors.","enum":[1000,1001,1002,1003,1004,1005,1006,2000,2001,2002]},"LoginRequest":{"type":"object","description":"The request data of a login request","required":["username","password"],"properties":{"username":{"type":"string","example":"user123"},"password":{"type":"string","example":"super-secure-password"}}},"ProductImages":{"type":"object","properties":{"image":{"type":"string","nullable":true}}},"SearchOutput":{"type":"object","required":["results"],"properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/SearchResult"}}}},"SearchResult":{"type":"object","required":["uuid","name","main_category"],"properties":{"uuid":{"type":"string","format":"uuid"},"name":{"type":"string"},"quantity":{"type":"string","nullable":true},"description":{"type":"string","nullable":true},"image":{"type":"string","nullable":true},"main_category":{"type":"string"}}}},"securitySchemes":{"session_cookie":{"type":"apiKey","in":"cookie","name":"id"}}}}